Move EmbeddedDialog to popup/ as PopupDialog
